efuses: define enums for AES mode and secure boot enable
Also spell out some more fields.
Note that we are unsure what the secure boot enable bits mean.
On the Sipeed MFST40 demo board with a screen attached, we get
1 for AES mode and 3 for secure boot enable. Per vendor code, the BL702
has an additional field for the secure boot _mode_. Something is odd here.
The vendor tool does define the bits for AWS mode, so we are certain.

Signed-off-by: Daniel Maslowski <info@orangecms.org>